+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added new BOM
 - Added [JSpecify](https://jspecify.dev/) and [NullAway](https://github.com/uber/nullaway)
 - Added projection lag metrics and outbox metrics using Micrometer
+- Added distributed HA projection lease
 
 ## 0.6.0
 - Added new Jackson module